Fix panic caused by mempool executor
Fixes #516
When the mempool executor was stopped, it was not being unsuscribed
from the network transaction topic, which is problematic after
it is restarted, since the network fails if we try to suscribe
and we are already suscribed.
