+++ mempool/tests/mod.rs
@@ -88,7 +88,7 @@ async fn send_txn_to_mempool(
 
     let timeout = tokio::time::Duration::from_secs(1);
     tokio::time::sleep(timeout).await;
-    mempool.stop_executor().await;
+    mempool.stop_executor_without_unsuscribe().await;
 }
 
 async fn multiple_start_stop_send(
@@ -126,7 +126,7 @@ async fn multiple_start_stop_send(
 
     let timeout = tokio::time::Duration::from_secs(2);
     tokio::time::sleep(timeout).await;
-    mempool.stop_executor().await;
+    mempool.stop_executor_without_unsuscribe().await;
 
     // Get the transactions from the mempool
     let obtained_txns = mempool.get_transactions_for_block(usize::MAX);
@@ -151,7 +151,7 @@ async fn multiple_start_stop_send(
     tokio::time::sleep(timeout).await;
 
     // Call stop again, nothing should happen.
-    mempool.stop_executor().await;
+    mempool.stop_executor_without_unsuscribe().await;
 
     // We should not obtain any, since the executor should not be running.
     let obtained_txns = mempool.get_transactions_for_block(usize::MAX);
@@ -189,7 +189,7 @@ async fn multiple_start_stop_send(
 
     let timeout = tokio::time::Duration::from_secs(2);
     tokio::time::sleep(timeout).await;
-    mempool.stop_executor().await;
+    mempool.stop_executor_without_unsuscribe().await;
 
     // Get the transactions from the mempool
     let obtained_txns = mempool.get_transactions_for_block(usize::MAX);
